Made sectrets test changes such that pxctl secrets login is not required. The px service update by enabling secrets is sufficient. Removed secrets tests from test_sanity and added it in test_updste.
