+++ test.sh
@@ -135,6 +135,10 @@ case $key in
     PX_IMAGE_ARGS="$2"
     shift
     ;;
+    --px-old-image)
+    PX_OLD_IMAGE_ARGS="$2"
+    shift
+    ;;
     -*)
     echo "Unknown option: $key"
     usage
@@ -263,6 +267,7 @@ docker run --rm \
     -e PYTEST_ARGS="$PYTEST_ARGS" \
     -e KVDB="$KVDB_ARGS" \
     -e PX_IMAGE="$PX_IMAGE_ARGS" \
+    -e PX_OLD_IMAGE="$PX_OLD_IMAGE_ARGS" \
     $FRAMEWORK_ARGS \
     -e STUB_UNIVERSE_URL="$STUB_UNIVERSE_URL" \
     -v $(pwd):$WORK_DIR \
